Move all Bash from script files into the README
This way, everything not automatically synchronized is explained in a
human-readable way, with links to sources. Or to put it another way,
everything not meant to be referred back to multiple times is in text.

I removed Yarn because [apparently][yarn] you use npm to install it, so
if I had it in here I'd probably just want to add code to `env.py` to
automatically synchronize global npm packages, and I don't want to do
that right now since I'm not really using Yarn at the moment anyway.

[yarn]: https://yarnpkg.com/getting-started/install#global-install