Sort random things in sub-sub-headings
The first two subsections are necessary to get the repo, and the
APT/Snap subsection is necessary to be able to do some of the other
things. The other things don't have dependencies on each other (yet...)
so putting them in alphabetical order will make it easy to insert and
find things in the future.